move utils into hpo dir
